+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add support for `orElse()` and `orElseEmpty()` on size-based gatherers to provide a non-exceptional output stream
 + Implement `everyNth()` to get every `n`<sup>th</sup> element from the stream
 + Implement `uniquelyOccurring()` to emit stream elements that occur a single time
++ Implement `takeUntil()` to take from a stream until a predicate is met, including the first element that matches the predicate
 
 ### 0.7.0
 + Use greedy integrators where possible (Fixes #57)